requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-filter==23.3
cryptography==41.0.4
dj-database-url==2.1.0